refactro the toolbar
